feat: implement update button
