+++ src/util/index.ts
@@ -5,7 +5,9 @@ export * from "./capitalize";
 export * from "./createCommand";
 export * from "./errorToResponse";
 export * from "./formatBytes";
+export * from "./getLiveEmbed";
 export * from "./loadMatchDay";
+export * from "./loadMatches";
 export * from "./normalizeError";
 export * from "./normalizeTeamName";
 export * from "./randomNumber";
